Puffin 'dist' file can now be read in, allowing the possibility of using the output from an external accelerator code in Puffin.
Several new files added. This will NOT read in the old style beam file.

It currently only works in 1D.

It is still to be decided how the initial conditions will be worked
out. The input file names are currently hard coded into the fortran files.
This is also the case for several other variables.

-Lawrence