New branch. Custom build for automatically matching the taper to compensate for the chirp, INCLUDING THE EFFECTS OF DISPERSION.
A linear taper can be used to counter the effects of the chirp.
However, when the chirp is large enough, the beam will also disperse
due to the chirp - the taper must be modified accordingly.
This custom version of puffin automatically calculates the taper
to nullify the effects of the chirp, based on the initial linear
chirp specified in the input files.

This allows one to isolate the other effect of the dispersion from
the chirp - the phase space squeezing and stretching.

-Lawrence