Fixing run time error.
The addition of the 'addChirp' subroutine had created a runtime
error in the case of using 2 or more beams. A leftover, redundant
array in CremLow.f90 was not being deallocated, causing an error
when trying to reallocate it when adding a chirp to a 2nd beam.

Now fixed. The redundant array has been removed.

-Lawrence