Now working for more general cases. Most sensible 1D cases will work - cases where the beam compresses will cause issues. This will be fixed soon.
Now rearranging data every 2 periods, with active buffers
of 4 wavelengths - this should be made variable.

Field and integrated power output not working from the distrubuted
data. Still using a global field array just to call the
data back to - but it is not being used in the calculations.
Should fix output to work with distributed data so the global
field can be removed completely.

-Lawrence