Adding back in strong focusing channel. Now controlled through vars sKBetaXSF and sKBetaYSF in main input file, specifying betatron wavenumber in x and y, respectively.
This strong focusing channel is probably highly unphysical -
but it will give a constant focusing channel in both transverse
dimensions. It is imposed *in addition to* the 'natural' focussing
channel from the undulator, which arises naturally and
self-consistently from the off-axis variation in the magnetic
undulator fields.

Focusfactor is no more!

-Lawrence