Now have flag 'qInitWrLat' to control the re-initialization of the writing steps at the start of each undulator module.
If this = .false. then the writes will be strictly spaced by the
options in the input file, regardless of the wiggler lattice layout.
For the moment, oddly, there is a full data write/dump at the end of the
FIRST module, and no data writing at the exit of the simulation
unless it's at an iWriteNthSteps step. This will be fixed soon!

If it = .true. then the writing steps will be re-initialized at the
start of each module - so you'll get every iWriteNthSteps written
from the start of the current module, etc. There will also be
a data write/dump at the end of each module.

-Lawrence