Adding input variable 'iMPsZ2PerWave', 'inmps1DGam', and 'qOneDCold'.
The latter 2 only work in the 1D case - when the field mesh is 1D.

qOneDCold forces the beam to be cold.

inmps1DGam overwrites iNumMPs in the beam file - essentially making
it redundant in the 1D case.

iMPsZ2PerWave works in both 3D and 1D modes. It allows the num of MPs
in z2 to be spec'd by the number of macroparticles per reference
wavelength. If present, it overwrites the value in iNumMPs.

-Lawrence